Se agregan mas custom errors y loggers

+++ src/controllers/orders.controller.js
@@ -1,10 +1,12 @@
-import CartsService from "../services/carts.service.js";
 import OrdersService from "../services/orders.service.js";
 import UsersService from "../services/users.service.js";
 import { v4 as uuidv4 } from 'uuid';
-import CartsController from "./carts.controller.js";
-import UserController from "./users.controller.js";
 import ProductsService from "../services/products.service.js";
+import { CustomError } from "../../utils/CustomErrors.js";
+import EnumsError from "../../utils/EnumsError.js";
+import { generatorOrderError, generatorOrderIdError, generatorUserIdError } from "../../utils/CauseMessageError.js";
+import { logger } from "../config/logger.js";
+import { Types  } from "mongoose";
 
 export default class OrdersController {
     static getAll(filter = {}) {
@@ -16,6 +18,7 @@ export default class OrdersController {
 
         const code = uuidv4();
         const userResult = await UsersService.getById(user);
+        logger.debug('UsersService.getById() finished successfully')
         const userId = user._id.toString();
 
         const currentUserWithCart = await userResult.populate('cart');
@@ -27,6 +30,7 @@ export default class OrdersController {
         const productsInCart = itemsInCart.products;
 
         const allProducts = await ProductsService.getAll();
+        logger.debug('ProductsService.getAll() finished successfully')
         const stockProducts = allProducts.filter(p => p.stock !== 0);
 
         const productsToPurchase = [];
@@ -53,16 +57,38 @@ export default class OrdersController {
         for (const item of productsToPurchase) {
             try {
                 const product = await ProductsService.getById(item.product);
+                logger.debug('ProductsService.getById() finished successfully')
 
                 if (product && product.price) {
                     const productTotalPrice = product.price * item.quantity;
                     totalPrice += productTotalPrice;
                 } else {
-                    console.error(`Error: No se pudo obtener información del producto ${item.product}`);
+                    logger.error(`Could not get product information. Product ID: ${item.product}`);
+                    throw new Error(`No se pudo obtener información del producto. ID del producto: ${item.product}`)
                 }
             } catch (error) {
-                console.error(`Error al obtener el producto ${item.product}: ${error.message}`);
+                logger.error(`Error getting product ${item.product}: ${error.message}`);
+                throw new Error(`Error al obtener el producto ${item.product}: ${error.message}`)
+
             }
+        }
+
+        if (!code ||
+            !userId ||
+            !productsToPurchase ||
+            totalPrice === 0
+        ) {
+            CustomError.create({
+                name: 'Invalid data order',
+                cause: generatorOrderError({
+                    code,
+                    user,
+                    products,
+                    total
+                }),
+                message: 'Error al crear una nueva orden',
+                code: EnumsError.BAD_REQUEST_ERROR,
+            })
         }
 
         const newOrder = {
@@ -72,42 +98,85 @@ export default class OrdersController {
             total: totalPrice,
         };
         const createdOrder = await OrdersService.create(newOrder);
+        logger.debug('OrdersService.create() finished successfully')
 
         const { orders } = userResult;
         orders.push(createdOrder);
         await UsersService.updateById(userId, { orders });
+        logger.debug('UsersService.updateById() finished successfully')
 
         return createdOrder;
     }
 
-    static async resolve(id, { status }) {
-        await OrdersService.updateById(id, { status });
-    }
-
     static async getById(id) {
-        const order = await OrdersService.getById(id);
-        if (!order) {
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid order id format',
+                cause: generatorOrderIdError(id),
+                message: 'Error al intentar obtener la orden por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingOrder = await OrdersService.getById(id);
+        logger.debug('OrdersService.getById() finished successfully')
+        if (!existingOrder) {
+            logger.error('Order not found')
             throw new Error(`Orden ${id} no encontrada`)
         }
-        return order
+        return existingOrder
     }
 
     static async updateById(id, data) {
-        await UserController.getById(id)
-        return UsersService.updateById(id, data);
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid order id format',
+                cause: generatorOrderIdError(id),
+                message: 'Error al intentar actualizar la orden por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingOrder = await OrdersService.getById(id);
+        logger.debug('OrdersService.getById() finished successfully')
+        if (!existingOrder) {
+            logger.error('Order not found')
+            throw new Error(`Orden ${id} no encontrada`)
+        }
+        return OrdersService.updateById(id, data);
     }
 
     static async deleteAllOrders(id) {
-        const user = await UsersService.getById(id);
-        if (!user) {
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid user id format',
+                cause: generatorUserIdError(id),
+                message: 'Error al intentar obtener el usuario por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingUser = await UsersService.getById(id);
+        logger.debug('UsersService.getById() finished successfully')
+        if (!existingUser) {
+            logger.error('User not found')
             throw new Error('Usuario no encontrado');
         }
         return UsersService.updateById(id, { orders: [] });
-
     }
 
     static async deleteById(id) {
-        await OrdersController.getById(id);
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid order id format',
+                cause: generatorOrderIdError(id),
+                message: 'Error al intentar eliminar la orden por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingOrder = await OrdersService.getById(id);
+        logger.debug('OrdersService.getById() finished successfully')
+        if (!existingOrder) {
+            logger.error('Order not found')
+            throw new Error(`Orden ${id} no encontrada`)
+        }
         return OrdersService.deleteById(id);
     }
 }
