+++ src/controllers/users.controller.js
@@ -1,8 +1,10 @@
 import UsersService from "../services/users.service.js";
 import { CustomError } from "../../utils/CustomErrors.js";
-import { generatorProductError } from "../../utils/CauseMessageError.js";
 import EnumsError from "../../utils/EnumsError.js";
+import { generatorUserError, generatorUserIdError } from "../../utils/CauseMessageError.js";
 import { logger } from "../config/logger.js";
+import { Types } from "mongoose";
+
 
 export default class UserController {
     static getAll(filter = {}) {
@@ -48,25 +50,67 @@ export default class UserController {
     }
 
     static async getById(id) {
-        const user = await UsersService.getById(id);
-        if (!user) {
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid user id format',
+                cause: generatorUserIdError(id),
+                message: 'Error al intentar obtener el usuario por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingUser = await UsersService.getById(id);
+        if (!existingUser) {
+            logger.error('User not found')
             throw new Error(`Usuario ${id} no encontrado`)
         }
-        return user
+        return existingUser
     }
 
     static async updateById(id, data) {
-        const existingUser = await UserController.getById(id);
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid user id format',
+                cause: generatorUserIdError(id),
+                message: 'Error al intentar obtener el usuario por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+         if (!data.first_name &&
+            !data.last_name &&
+            !data.age) {
+            CustomError.create({
+                name: 'Invalid data user',
+                cause: generatorUserError({
+                    first_name,
+                    last_name,
+                    age
+                }),
+                message: `Error al actualizar el usuario ${id}`,
+                code: EnumsError.BAD_REQUEST_ERROR,
+            })
+        }
+        const existingUser = await UsersService.getById(id);
         if (!existingUser) {
             logger.error('User not found')
             throw new Error('Usuario no encontrado');
         }
-        logger.info(`User successfully updated: ${id}`)
         return UsersService.updateById(id, data);
     }
 
     static async deleteById(id) {
-        await UserController.getById(id)
+        if (!Types.ObjectId.isValid(id)) {
+            CustomError.create({
+                name: 'Invalid user id format',
+                cause: generatorUserIdError(id),
+                message: 'Error al intentar obtener el usuario por su id',
+                code: EnumsError.INVALID_PARAMS_ERROR
+            });
+        }
+        const existingUser = await UsersService.getById(id);
+        if (!existingUser) {
+            logger.error('User not found')
+            throw new Error('Usuario no encontrado');
+        }
         return UsersService.deleteById(id);
     }
 }
